Move order logic to a separate file
