+++ .storybook/preview.js
@@ -6,37 +6,14 @@ import { withKnobs } from '@storybook/addon-knobs'
 import i18n from '@/i18n'
 
 import viewports from './viewport'
+import order from './order'
 
 import '@creativecommons/vocabulary/css/vocabulary.css'
-
-const order = {
-  'Vue Vocabulary': [
-    'Introduction',
-    'Contribution',
-    'Usage'
-  ]
-}
-const families = Object.keys(order)
 
 addParameters({
   options: {
     showRoots: true,
-    storySort: ([, one], [, two]) => {
-      if (one.kind === two.kind) {
-        return 0 // Sort stories in a component as defined in the MDX file
-      }
-      const [famOne, componentOne] = one.kind.split('/')
-      const [famTwo, componentTwo] = two.kind.split('/')
-      if (famOne === famTwo) {
-        if (order[famOne].length) {
-          return order[famOne].indexOf(componentOne) - order[famOne].indexOf(componentTwo)
-        } else {
-          return componentOne.localeCompare(componentTwo) // Sort components in a family in alphabetical order
-        }
-      } else {
-        return families.indexOf(famOne) - families.indexOf(famTwo) // Sort families according to defined order
-      }
-    }
+    storySort: order
   },
   backgrounds: [
     { name: 'canvas', value: '#f5f5f5', default: true },
